fix(#2827): C1/C5/C6 — cardinality rustdoc, exported loss signals, guard holes
C1 the last twelve-series claim (REPEAT_BUCKET rustdoc). C5 dropped_accesses and
sampling_floor are now catalogued, annotated, exported with dedicated otel arms and
asserted in the capture-harness test, so an operator reading dashboards alone can
tell a lossy or floored window from a clean one. C6: the catalog guard scans BOTH
otel files (instrument construction moved to otel_instruments.rs for headroom, so
scanning one would have blinded it), its parser no longer truncates on a semicolon
inside a value, a named test pins the wrapped-declaration parse, the emission-site
census gains bti_lookup_memo.rs, the histogram capture test uses the B7 window hook,
tasks.md marks its superseded lines at the point of use, and the resolver carries the
load-bearing 'must not be used for the decision' clause.

Ranking formula untouched pending the owner ruling.
