+++ cqlite-core/src/observability/operator_docs_annotations.rs
@@ -92,6 +92,24 @@ pub(super) const ANNOTATIONS: &[MetricDoc] = &[
         round_item: "—",
     },
     MetricDoc {
+        name: catalog::READ_PARTITION_ACCESS_DROPPED,
+        kind: MetricKind::Counter,
+        unit: catalog::unit::DIMENSIONLESS,
+        summary: "Accesses the access-distribution probe could not seat in its fixed counting table (#2827).",
+        attributes: &[],
+        interpretation: "Zero on a healthy window. Non-zero INVALIDATES the window: only keys not already in the table can be dropped, so a loss suppresses the singleton bucket and overstates concentration — the cache-sizing procedure refuses such a window.",
+        round_item: "—",
+    },
+    MetricDoc {
+        name: catalog::READ_PARTITION_ACCESS_SAMPLING_FLOOR,
+        kind: MetricKind::Gauge,
+        unit: catalog::unit::DIMENSIONLESS,
+        summary: "1 when the last closed probe window reached its sampling-prefix cap, 0 otherwise (#2827).",
+        attributes: &[],
+        interpretation: "0 is healthy. 1 means the window is a ~1-in-a-million sample and is statistically worthless; the cache-sizing procedure refuses it. Read alongside sample_denominator.",
+        round_item: "—",
+    },
+    MetricDoc {
         name: catalog::READ_BLOOM_CHECKS,
         kind: MetricKind::Counter,
         unit: catalog::unit::DIMENSIONLESS,
