+++ cqlite-core/src/observability/otel.rs
@@ -13,7 +13,8 @@ use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::OnceLock;
 use std::time::Duration;
 
-use opentelemetry::metrics::{Counter, Gauge, Histogram, Meter};
+use super::otel_instruments::instruments;
+use opentelemetry::metrics::Meter;
 use opentelemetry::trace::TracerProvider as _;
 use opentelemetry::{global, KeyValue};
 use opentelemetry_otlp::{MetricExporter, Protocol, SpanExporter, WithExportConfig};
@@ -271,494 +272,9 @@ where
 }
 
 /// The global CQLite [`Meter`], cached after first use.
-fn meter() -> &'static Meter {
+pub(super) fn meter() -> &'static Meter {
     static METER: OnceLock<Meter> = OnceLock::new();
     METER.get_or_init(|| global::meter(SCOPE))
-}
-
-/// Lazily-built, cached instruments for every catalog metric. Building an
-/// instrument on each record call is wasteful (re-registration overhead and
-/// possible duplicate-instrument churn), so all catalog instruments are
-/// constructed once and reused. Non-catalog names fall back to an ad-hoc
-/// instrument so call sites never silently drop data.
-struct Instruments {
-    read_rows: Counter<u64>,
-    read_bytes: Counter<u64>,
-    read_partitions: Counter<u64>,
-    read_partition_lookup: Counter<u64>,
-    read_bloom_checks: Counter<u64>,
-    read_scan_window_refill: Counter<u64>,
-    read_sstables_pruned: Counter<u64>,
-    read_bloom_false_negatives: Counter<u64>,
-    read_bti_rows_root_rejected: Counter<u64>,
-    read_partition_access_distinct: Counter<u64>,
-    read_partition_access_accesses: Counter<u64>,
-    read_partition_access_bytes: Counter<u64>,
-    read_partition_access_sample_denominator: Gauge<i64>,
-    merge_rows_in: Counter<u64>,
-    merge_rows_out: Counter<u64>,
-    query_degraded_path: Counter<u64>,
-    index_parses_total: Counter<u64>,
-    index_interval_parses_total: Counter<u64>,
-    storage_open_sstables: Counter<u64>,
-    storage_open_bytes: Counter<u64>,
-    storage_open_tables: Counter<u64>,
-    query_rows: Counter<u64>,
-    query_rows_scanned: Counter<u64>,
-    errors_total: Counter<u64>,
-    write_mutations: Counter<u64>,
-    flush_rows: Counter<u64>,
-    flush_bytes: Counter<u64>,
-    flush_sstables: Counter<u64>,
-    write_partitions: Counter<u64>,
-    write_bytes: Counter<u64>,
-    compaction_rows_merged: Counter<u64>,
-    compaction_bytes_written: Counter<u64>,
-    compaction_sstables_in: Counter<u64>,
-    compaction_sstables_out: Counter<u64>,
-    compaction_tombstones_purged: Counter<u64>,
-    compaction_tombstones_suppressed: Counter<u64>,
-    compaction_tombstones_emitted: Counter<u64>,
-    rpc_requests: Counter<u64>,
-    rpc_rows: Counter<u64>,
-    rpc_bytes: Counter<u64>,
-    warm_cache_hits: Counter<u64>,
-    warm_cache_misses: Counter<u64>,
-    warm_cache_evicts: Counter<u64>,
-    warm_cache_refresh: Counter<u64>,
-    flight_admission_rejected_total: Counter<u64>,
-    read_duration: Histogram<f64>,
-    query_duration: Histogram<f64>,
-    compaction_duration: Histogram<f64>,
-    wal_sync_duration: Histogram<f64>,
-    flush_duration: Histogram<f64>,
-    compression_ratio: Histogram<f64>,
-    compaction_finalize_duration: Histogram<f64>,
-    compaction_budget_requested: Histogram<f64>,
-    compaction_budget_consumed: Histogram<f64>,
-    rpc_duration: Histogram<f64>,
-    rpc_phase_duration: Histogram<f64>,
-    flight_admission_wait_seconds: Histogram<f64>,
-    sstables_open: Gauge<i64>,
-    memtable_size_bytes: Gauge<i64>,
-    memtable_rows: Gauge<i64>,
-    compaction_lag: Gauge<i64>,
-    rpc_in_flight: Gauge<i64>,
-    rpc_phase_active: Gauge<i64>,
-    merge_producer_threads: Gauge<i64>,
-    flight_admission_limit: Gauge<i64>,
-    flight_admission_in_use: Gauge<i64>,
-    flight_admission_waiting: Gauge<i64>,
-    // Saturation instrumentation (#2419, WS2 of epic #2313).
-    merge_egress_channel_depth: Gauge<i64>,
-    // Adaptive egress-budget concurrency gauge (#2765).
-    merge_active_merges: Gauge<i64>,
-    proc_threads: Gauge<i64>,
-    proc_fds: Gauge<i64>,
-    proc_rss_bytes: Gauge<i64>,
-    flight_blocking_tasks_in_use: Gauge<i64>,
-    // Flight table-visibility gauges (#2684).
-    flight_tables_discovered: Gauge<i64>,
-    flight_warm_tables: Gauge<i64>,
-}
-
-fn instruments() -> &'static Instruments {
-    static INSTRUMENTS: OnceLock<Instruments> = OnceLock::new();
-    INSTRUMENTS.get_or_init(|| {
-        let m = meter();
-        Instruments {
-            read_rows: m
-                .u64_counter(catalog::READ_ROWS)
-                .with_unit(catalog::unit::ROWS)
-                .with_description("Total rows materialised by the read path.")
-                .build(),
-            read_bytes: m
-                .u64_counter(catalog::READ_BYTES)
-                .with_unit(catalog::unit::BYTES)
-                .with_description("Total bytes read from Data.db (post-decompression).")
-                .build(),
-            read_partitions: m
-                .u64_counter(catalog::READ_PARTITIONS)
-                .with_unit(catalog::unit::PARTITIONS)
-                .with_description("Total partitions scanned.")
-                .build(),
-            read_partition_lookup: m
-                .u64_counter(catalog::READ_PARTITION_LOOKUP)
-                .with_unit(catalog::unit::DIMENSIONLESS)
-                .with_description("Total partition point lookups, keyed by {result, lookup_route, format}.")
-                .build(),
-            read_bloom_checks: m
-                .u64_counter(catalog::READ_BLOOM_CHECKS)
-                .with_unit(catalog::unit::DIMENSIONLESS)
-                .with_description("Total bloom/BTI-trie presence checks, keyed by {result}.")
-                .build(),
-            read_scan_window_refill: m
-                .u64_counter(catalog::READ_SCAN_WINDOW_REFILL)
-                .with_unit(catalog::unit::DIMENSIONLESS)
-                .with_description("Windowed scan refills at compression-chunk boundaries.")
-                .build(),
-            read_sstables_pruned: m
-                .u64_counter(catalog::READ_SSTABLES_PRUNED)
-                .with_unit(catalog::unit::SSTABLES)
-                .with_description(
-                    "SSTables skipped by a presence-oracle negative, keyed by {format}.",
-                )
-                .build(),
-            read_bloom_false_negatives: m
-                .u64_counter(catalog::READ_BLOOM_FALSE_NEGATIVES)
-                .with_unit(catalog::unit::DIMENSIONLESS)
-                .with_description(
-                    "Opt-in presence-oracle false negatives (soundness alarm), keyed by {format}.",
-                )
-                .build(),
-            read_bti_rows_root_rejected: m
-                .u64_counter(catalog::READ_BTI_ROWS_ROOT_REJECTED)
-                .with_unit(catalog::unit::PARTITIONS)
-                .with_description(
-                    "Clustering reads that decoded a whole BTI partition because its Rows.db \
-                     row-index root failed validation, keyed by {reason} (#3002).",
-                )
-                .build(),
-            read_partition_access_distinct: m
-                .u64_counter(catalog::READ_PARTITION_ACCESS_DISTINCT_PARTITIONS)
-                .with_unit(catalog::unit::PARTITIONS)
-                .with_description("Distinct partitions per repeat-access bucket (#2827).")
-                .build(),
-            read_partition_access_accesses: m
-                .u64_counter(catalog::READ_PARTITION_ACCESS_ACCESSES)
-                .with_unit(catalog::unit::DIMENSIONLESS)
-                .with_description("Accesses per repeat-access bucket (#2827).")
-                .build(),
-            read_partition_access_bytes: m
-                .u64_counter(catalog::READ_PARTITION_ACCESS_BYTES)
-                .with_unit(catalog::unit::BYTES)
-                .with_description("Distinct-partition on-disk bytes per bucket (#2827).")
-                .build(),
-            read_partition_access_sample_denominator: m
-                .i64_gauge(catalog::READ_PARTITION_ACCESS_SAMPLE_DENOMINATOR)
-                .with_unit(catalog::unit::DIMENSIONLESS)
-                .with_description("Probe sampling scale at window close; 1 = census.")
-                .build(),
-            merge_rows_in: m
-                .u64_counter(catalog::MERGE_ROWS_IN)
-                .with_unit(catalog::unit::ROWS)
-                .with_description("Rows consumed at the k-way merge reconcile boundary.")
-                .build(),
-            merge_rows_out: m
-                .u64_counter(catalog::MERGE_ROWS_OUT)
-                .with_unit(catalog::unit::ROWS)
-                .with_description("Rows emitted by the k-way merge reconcile boundary.")
-                .build(),
-            query_degraded_path: m
-                .u64_counter(catalog::QUERY_DEGRADED_PATH)
-                .with_unit(catalog::unit::DIMENSIONLESS)
-                .with_description(
-                    "SELECTs taking a soundness fallback, keyed by {fallback_reason}.",
-                )
-                .build(),
-            index_parses_total: m
-                .u64_counter(catalog::INDEX_PARSES_TOTAL)
-                .with_unit(catalog::unit::DIMENSIONLESS)
-                .with_description("Full Index.db partition-index parses (#2383 spin probe).")
-                .build(),
-            index_interval_parses_total: m
-                .u64_counter(catalog::INDEX_INTERVAL_PARSES_TOTAL)
-                .with_unit(catalog::unit::DIMENSIONLESS)
-                .with_description(
-                    "Bounded Summary-guided Index.db interval parses, per point lookup (issue #2412).",
-                )
-                .build(),
-            storage_open_sstables: m
-                .u64_counter(catalog::STORAGE_OPEN_SSTABLES)
-                .with_unit(catalog::unit::SSTABLES)
-                .with_description("SSTables discovered and opened, summed across opens.")
-                .build(),
-            storage_open_bytes: m
-                .u64_counter(catalog::STORAGE_OPEN_BYTES)
-                .with_unit(catalog::unit::BYTES)
-                .with_description("On-disk Data.db bytes across SSTables discovered at open.")
-                .build(),
-            storage_open_tables: m
-                .u64_counter(catalog::STORAGE_OPEN_TABLES)
-                .with_unit(catalog::unit::DIMENSIONLESS)
-                .with_description("Logical tables represented by SSTables discovered at open.")
-                .build(),
-            query_rows: m
-                .u64_counter(catalog::QUERY_ROWS)
-                .with_unit(catalog::unit::ROWS)
-                .with_description("Total rows returned to callers by the query engine.")
-                .build(),
-            query_rows_scanned: m
-                .u64_counter(catalog::QUERY_ROWS_SCANNED)
-                .with_unit(catalog::unit::ROWS)
-                .with_description("Rows examined by SELECT scan before filtering/projection.")
-                .build(),
-            errors_total: m
-                .u64_counter(catalog::ERRORS_TOTAL)
-                .with_unit(catalog::unit::ERRORS)
-                .with_description("Total errors observed, keyed by bounded {category, subsystem}.")
-                .build(),
-            write_mutations: m
-                .u64_counter(catalog::WRITE_MUTATIONS)
-                .with_unit(catalog::unit::ROWS)
-                .with_description("Mutations accepted by the write path.")
-                .build(),
-            flush_rows: m
-                .u64_counter(catalog::FLUSH_ROWS)
-                .with_unit(catalog::unit::ROWS)
-                .with_description("Rows flushed from memtable to L0 SSTables.")
-                .build(),
-            flush_bytes: m
-                .u64_counter(catalog::FLUSH_BYTES)
-                .with_unit(catalog::unit::BYTES)
-                .with_description("Data.db bytes produced by memtable flushes.")
-                .build(),
-            flush_sstables: m
-                .u64_counter(catalog::FLUSH_SSTABLES)
-                .with_unit(catalog::unit::SSTABLES)
-                .with_description("L0 SSTables created by memtable flushes.")
-                .build(),
-            write_partitions: m
-                .u64_counter(catalog::WRITE_PARTITIONS)
-                .with_unit(catalog::unit::PARTITIONS)
-                .with_description("Partitions written by the SSTable writer.")
-                .build(),
-            write_bytes: m
-                .u64_counter(catalog::WRITE_BYTES)
-                .with_unit(catalog::unit::BYTES)
-                .with_description("Data.db bytes produced by the SSTable writer.")
-                .build(),
-            compaction_rows_merged: m
-                .u64_counter(catalog::COMPACTION_ROWS_MERGED)
-                .with_unit(catalog::unit::ROWS)
-                .with_description("Rows emitted by compaction merge.")
-                .build(),
-            compaction_bytes_written: m
-                .u64_counter(catalog::COMPACTION_BYTES_WRITTEN)
-                .with_unit(catalog::unit::BYTES)
-                .with_description("Bytes written to compaction output SSTables.")
-                .build(),
-            compaction_sstables_in: m
-                .u64_counter(catalog::COMPACTION_SSTABLES_IN)
-                .with_unit(catalog::unit::SSTABLES)
-                .with_description("Input SSTables consumed by compactions.")
-                .build(),
-            compaction_sstables_out: m
-                .u64_counter(catalog::COMPACTION_SSTABLES_OUT)
-                .with_unit(catalog::unit::SSTABLES)
-                .with_description("Output SSTables produced by compactions.")
-                .build(),
-            compaction_tombstones_purged: m
-                .u64_counter(catalog::COMPACTION_TOMBSTONES_PURGED)
-                .with_unit(catalog::unit::TOMBSTONES)
-                .with_description("Tombstones genuinely purged during compaction.")
-                .build(),
-            compaction_tombstones_suppressed: m
-                .u64_counter(catalog::COMPACTION_TOMBSTONES_SUPPRESSED)
-                .with_unit(catalog::unit::TOMBSTONES)
-                .with_description("Live cells/rows shadowed by a tombstone during reconciliation.")
-                .build(),
-            compaction_tombstones_emitted: m
-                .u64_counter(catalog::COMPACTION_TOMBSTONES_EMITTED)
-                .with_unit(catalog::unit::TOMBSTONES)
-                .with_description("Tombstone markers retained into the merge output.")
-                .build(),
-            rpc_requests: m
-                .u64_counter(catalog::RPC_REQUESTS)
-                .with_unit(catalog::unit::DIMENSIONLESS)
-                .with_description("Arrow Flight RPC requests served.")
-                .build(),
-            rpc_rows: m
-                .u64_counter(catalog::RPC_ROWS)
-                .with_unit(catalog::unit::ROWS)
-                .with_description("Rows returned to Flight clients.")
-                .build(),
-            rpc_bytes: m
-                .u64_counter(catalog::RPC_BYTES)
-                .with_unit(catalog::unit::BYTES)
-                .with_description("Record-batch payload bytes streamed to Flight clients.")
-                .build(),
-            warm_cache_hits: m
-                .u64_counter(catalog::WARM_CACHE_HITS)
-                .with_unit(catalog::unit::DIMENSIONLESS)
-                .with_description("Flight warm-handle cache hits (#2310).")
-                .build(),
-            warm_cache_misses: m
-                .u64_counter(catalog::WARM_CACHE_MISSES)
-                .with_unit(catalog::unit::DIMENSIONLESS)
-                .with_description("Flight warm-handle cache misses (#2310).")
-                .build(),
-            warm_cache_evicts: m
-                .u64_counter(catalog::WARM_CACHE_EVICTS)
-                .with_unit(catalog::unit::DIMENSIONLESS)
-                .with_description("Warm generations evicted (LRU / removed on disk) (#2310).")
-                .build(),
-            warm_cache_refresh: m
-                .u64_counter(catalog::WARM_CACHE_REFRESH)
-                .with_unit(catalog::unit::DIMENSIONLESS)
-                .with_description("Warm-handle refresh outcomes, keyed by {refresh_outcome} (#2310).")
-                .build(),
-            flight_admission_rejected_total: m
-                .u64_counter(catalog::FLIGHT_ADMISSION_REJECTED_TOTAL)
-                .with_unit(catalog::unit::DIMENSIONLESS)
-                .with_description("do_get requests rejected on admission timeout (#2420).")
-                .build(),
-            read_duration: m
-                .f64_histogram(catalog::READ_DURATION)
-                .with_unit(catalog::unit::SECONDS)
-                .with_description("Single read/scan operation duration in seconds.")
-                .build(),
-            query_duration: m
-                .f64_histogram(catalog::QUERY_DURATION)
-                .with_unit(catalog::unit::SECONDS)
-                .with_description("End-to-end query execution duration in seconds.")
-                .build(),
-            compaction_duration: m
-                .f64_histogram(catalog::COMPACTION_DURATION)
-                .with_unit(catalog::unit::SECONDS)
-                .with_description("Compaction run duration in seconds.")
-                .build(),
-            wal_sync_duration: m
-                .f64_histogram(catalog::WAL_SYNC_DURATION)
-                .with_unit(catalog::unit::SECONDS)
-                .with_description("WAL fsync duration in seconds.")
-                .build(),
-            flush_duration: m
-                .f64_histogram(catalog::FLUSH_DURATION)
-                .with_unit(catalog::unit::SECONDS)
-                .with_description("Memtable-to-SSTable flush duration in seconds.")
-                .build(),
-            compression_ratio: m
-                .f64_histogram(catalog::COMPRESSION_RATIO)
-                .with_unit(catalog::unit::DIMENSIONLESS)
-                .with_description("Per-chunk compression ratio.")
-                .build(),
-            compaction_finalize_duration: m
-                .f64_histogram(catalog::COMPACTION_FINALIZE_DURATION)
-                .with_unit(catalog::unit::SECONDS)
-                .with_description("Compaction finalize duration in seconds.")
-                .build(),
-            compaction_budget_requested: m
-                .f64_histogram(catalog::COMPACTION_BUDGET_REQUESTED)
-                .with_unit(catalog::unit::SECONDS)
-                .with_description("Maintenance budget requested in seconds.")
-                .build(),
-            compaction_budget_consumed: m
-                .f64_histogram(catalog::COMPACTION_BUDGET_CONSUMED)
-                .with_unit(catalog::unit::SECONDS)
-                .with_description("Maintenance budget consumed in seconds.")
-                .build(),
-            rpc_duration: m
-                .f64_histogram(catalog::RPC_DURATION)
-                .with_unit(catalog::unit::SECONDS)
-                .with_description("Arrow Flight RPC handler duration in seconds.")
-                .build(),
-            rpc_phase_duration: m
-                .f64_histogram(catalog::RPC_PHASE_DURATION)
-                .with_unit(catalog::unit::SECONDS)
-                .with_description("do_get per-phase duration in seconds (#2162).")
-                .build(),
-            flight_admission_wait_seconds: m
-                .f64_histogram(catalog::FLIGHT_ADMISSION_WAIT_SECONDS)
-                .with_unit(catalog::unit::SECONDS)
-                .with_description("do_get admission acquire wait time in seconds (#2420).")
-                .build(),
-            sstables_open: m
-                .i64_gauge(catalog::SSTABLES_OPEN)
-                .with_unit(catalog::unit::SSTABLES)
-                .with_description("Number of SSTables currently held open.")
-                .build(),
-            memtable_size_bytes: m
-                .i64_gauge(catalog::MEMTABLE_SIZE_BYTES)
-                .with_unit(catalog::unit::BYTES)
-                .with_description("Approximate active memtable size in bytes.")
-                .build(),
-            memtable_rows: m
-                .i64_gauge(catalog::MEMTABLE_ROWS)
-                .with_unit(catalog::unit::ROWS)
-                .with_description("Rows currently buffered in the active memtable.")
-                .build(),
-            compaction_lag: m
-                .i64_gauge(catalog::COMPACTION_LAG)
-                .with_unit(catalog::unit::SSTABLES)
-                .with_description("Current L0 SSTables pending compaction.")
-                .build(),
-            rpc_in_flight: m
-                .i64_gauge(catalog::RPC_IN_FLIGHT)
-                .with_unit(catalog::unit::DIMENSIONLESS)
-                .with_description("Arrow Flight RPCs currently being handled.")
-                .build(),
-            rpc_phase_active: m
-                .i64_gauge(catalog::RPC_PHASE_ACTIVE)
-                .with_unit(catalog::unit::DIMENSIONLESS)
-                .with_description("do_get phase currently executing (#2361).")
-                .build(),
-            merge_producer_threads: m
-                .i64_gauge(catalog::MERGE_PRODUCER_THREADS)
-                .with_unit(catalog::unit::THREADS)
-                .with_description("Live k-way merge producer threads (#2316).")
-                .build(),
-            flight_admission_limit: m
-                .i64_gauge(catalog::FLIGHT_ADMISSION_LIMIT)
-                .with_unit(catalog::unit::DIMENSIONLESS)
-                .with_description("Configured do_get admission ceiling K (#2420).")
-                .build(),
-            flight_admission_in_use: m
-                .i64_gauge(catalog::FLIGHT_ADMISSION_IN_USE)
-                .with_unit(catalog::unit::DIMENSIONLESS)
-                .with_description("do_get admission permits currently held (#2420).")
-                .build(),
-            flight_admission_waiting: m
-                .i64_gauge(catalog::FLIGHT_ADMISSION_WAITING)
-                .with_unit(catalog::unit::DIMENSIONLESS)
-                .with_description("do_get requests parked waiting for an admission permit (#2420).")
-                .build(),
-            merge_egress_channel_depth: m
-                .i64_gauge(catalog::MERGE_EGRESS_CHANNEL_DEPTH)
-                .with_unit(catalog::unit::ENTRIES)
-                .with_description("Live occupancy of the bounded merge egress sync_channel (#2419).")
-                .build(),
-            merge_active_merges: m
-                .i64_gauge(catalog::MERGE_ACTIVE_MERGES)
-                .with_unit(catalog::unit::MERGES)
-                .with_description(
-                    "Live concurrent k-way merges — divisor of the adaptive egress budget (#2765).",
-                )
-                .build(),
-            proc_threads: m
-                .i64_gauge(catalog::PROC_THREADS)
-                .with_unit(catalog::unit::THREADS)
-                .with_description("Process OS thread count (/proc/self/task, Linux) (#2419).")
-                .build(),
-            proc_fds: m
-                .i64_gauge(catalog::PROC_FDS)
-                .with_unit(catalog::unit::FDS)
-                .with_description("Process open fd count (/proc/self/fd, Linux) (#2419).")
-                .build(),
-            proc_rss_bytes: m
-                .i64_gauge(catalog::PROC_RSS_BYTES)
-                .with_unit(catalog::unit::BYTES)
-                .with_description("Process resident set size (/proc/self/status VmRSS, Linux) (#2419).")
-                .build(),
-            flight_blocking_tasks_in_use: m
-                .i64_gauge(catalog::FLIGHT_BLOCKING_TASKS_IN_USE)
-                .with_unit(catalog::unit::THREADS)
-                .with_description("Flight spawn_blocking tasks currently outstanding (#2419).")
-                .build(),
-            flight_tables_discovered: m
-                .i64_gauge(catalog::FLIGHT_TABLES_DISCOVERED)
-                .with_unit(catalog::unit::ENTRIES)
-                .with_description(
-                    "Table dirs visible under --data-dir, sampled by readdir on the ~2s tick (#2684).",
-                )
-                .build(),
-            flight_warm_tables: m
-                .i64_gauge(catalog::FLIGHT_WARM_TABLES)
-                .with_unit(catalog::unit::ENTRIES)
-                .with_description("Tables with a live warm reader set in the registry (#2684).")
-                .build(),
-        }
-    })
 }
 
 /// Add to a u64 counter identified by a catalog name.
@@ -781,6 +297,7 @@ pub(crate) fn add_counter(name: &'static str, value: u64, attributes: &[KeyValue
         catalog::READ_PARTITION_ACCESS_DISTINCT_PARTITIONS => &i.read_partition_access_distinct,
         catalog::READ_PARTITION_ACCESS_ACCESSES => &i.read_partition_access_accesses,
         catalog::READ_PARTITION_ACCESS_BYTES => &i.read_partition_access_bytes,
+        catalog::READ_PARTITION_ACCESS_DROPPED => &i.read_partition_access_dropped,
         catalog::MERGE_ROWS_IN => &i.merge_rows_in,
         catalog::MERGE_ROWS_OUT => &i.merge_rows_out,
         catalog::QUERY_DEGRADED_PATH => &i.query_degraded_path,
@@ -856,6 +373,7 @@ pub(crate) fn record_gauge(name: &'static str, value: i64, attributes: &[KeyValu
         catalog::READ_PARTITION_ACCESS_SAMPLE_DENOMINATOR => {
             &i.read_partition_access_sample_denominator
         }
+        catalog::READ_PARTITION_ACCESS_SAMPLING_FLOOR => &i.read_partition_access_sampling_floor,
         catalog::MEMTABLE_SIZE_BYTES => &i.memtable_size_bytes,
         catalog::MEMTABLE_ROWS => &i.memtable_rows,
         catalog::COMPACTION_LAG => &i.compaction_lag,
