+++ docs/reports/flight-metrics-reference.md
@@ -6,7 +6,7 @@ Operator-facing reference for every `cqlite.*` instrument CQLite emits over Arro
 
 Related: the Flight/Trino operator docs (`docs/flight-trino/`) and the round scoreboard template (issue #2399) link back to the entries here.
 
-Total instruments: **81**.
+Total instruments: **83**.
 
 ## All instruments
 
@@ -66,7 +66,9 @@ Total instruments: **81**.
 | `cqlite.read.partition_access.accesses` | counter | `1` | `cqlite.read.repeat_bucket` | Accesses attributable to each repeat-access bucket over one closed probe window (#2827) — the sum of the repeat counts of the partitions in that bucket. | accesses minus distinct_partitions is the number of reads a cache holding that bucket could have served; equal values mean every access was a first touch (nothing to cache). |
 | `cqlite.read.partition_access.bytes` | counter | `By` | `cqlite.read.repeat_bucket` | Distinct-partition on-disk bytes per repeat-access bucket over one closed probe window (#2827), measured as each partition's successor gap; a partition read ten times counts its bytes once. | Non-zero on any window whose extents were measurable; uncompressed on-disk bytes, which is the correct input for a decoded-size multiplier. Zero only when every access reported size_source=unavailable. Never an estimate. |
 | `cqlite.read.partition_access.distinct_partitions` | counter | `{partition}` | `cqlite.read.repeat_bucket`<br>`cqlite.read.size_source` | Distinct partitions per repeat-access bucket over one closed window of the default-OFF access-distribution probe (#2827) — the workload's concentration shape, with no per-key label. | Absent unless an operator enabled CQLITE_PARTITION_ACCESS_PROBE. Mass in bucket 1 is a uniform, cache-hostile pattern; mass in 9-16/17+ is a hot set. size_source=successor_gap is a measured on-disk extent; a non-zero size_source=unavailable share means the byte total beside it is incomplete and the cache-sizing procedure refuses the window. |
+| `cqlite.read.partition_access.dropped_accesses` | counter | `1` | _(none)_ | Accesses the access-distribution probe could not seat in its fixed counting table (#2827). | Zero on a healthy window. Non-zero INVALIDATES the window: only keys not already in the table can be dropped, so a loss suppresses the singleton bucket and overstates concentration — the cache-sizing procedure refuses such a window. |
 | `cqlite.read.partition_access.sample_denominator` | gauge | `1` | _(none)_ | Sampling scale in force when the probe window closed (#2827): 1 is a census, 2^k means the recorder downsampled k times to stay inside its fixed 3 MiB table. | 1 is ideal. A large value means the window touched far more distinct partitions than the table holds; bucket fractions stay unbiased but absolute counts must be scaled by it. |
+| `cqlite.read.partition_access.sampling_floor` | gauge | `1` | _(none)_ | 1 when the last closed probe window reached its sampling-prefix cap, 0 otherwise (#2827). | 0 is healthy. 1 means the window is a ~1-in-a-million sample and is statistically worthless; the cache-sizing procedure refuses it. Read alongside sample_denominator. |
 | `cqlite.read.partition_lookup.total` | counter | `1` | `cqlite.result`<br>`cqlite.read.lookup_route`<br>`cqlite.sstable.format` | Partition point lookups attempted, tagged hit/miss so a dashboard computes the hit ratio from one series. | A healthy point-read workload is dominated by hits; a miss-heavy ratio means keys are absent or mis-routed. |
 | `cqlite.read.partitions` | counter | `{partition}` | `cqlite.sstable.format` | Total partitions scanned by the read path. | Rising in step with read.rows is healthy; many partitions for few rows indicates a full scan. |
 | `cqlite.read.rows` | counter | `{row}` | `cqlite.sstable.format` | Total rows materialised by the read path (climbs incrementally during a long Flight merge scan). | Steadily rising under load is healthy; flat while a scan is in flight suggests a stall. |
