+++ openspec/changes/keyed-access-distribution-probe/specs/partition-access-distribution/spec.md
@@ -97,6 +97,13 @@ The scale in force at window close SHALL be published as
 `cqlite.read.partition_access.sample_denominator`, so a consumer can distinguish a census
 (`denominator = 1`) from a sample.
 
+Accesses the recorder could NOT seat in the table SHALL be counted and published
+(`cqlite.read.partition_access.dropped_accesses`), and whether the window reached the prefix cap
+SHALL be published (`cqlite.read.partition_access.sampling_floor`), so a consumer reading the
+emitted series ALONE can tell a lossy or floored window from a clean one. A probe-cluster failure
+to seat a key SHALL widen the sample rather than drop the key: only keys not already in the table
+can fail to seat, so dropping them would suppress the singleton bucket and OVERSTATE concentration.
+
 The prefix width SHALL be capped, and a window that reaches the cap SHALL be marked non-census so the
 decision procedure refuses it (see *A committed decision procedure…*).
 
@@ -340,8 +347,10 @@ It SHALL state, at minimum:
 - that the result is an **upper bound**, so a low value is a sound **no-go** while a high value is
   necessary but not sufficient for a "go";
 - the **refusal conditions**, each of which yields *no answer* rather than a default verdict: a
-  non-zero `unavailable` fraction, a non-census window at the sampling floor, a window below a stated
-  minimum access count, and a window generated by synthetic or self-generated load;
+  non-zero `unavailable` fraction, accesses the recorder could not seat at all, a window at the
+  sampling floor, a window that is a SAMPLE rather than a census (its per-bucket bytes are
+  sample-domain totals, so filling a real budget against them overstates what fits), a window below
+  a stated minimum access count, and a window generated by synthetic or self-generated load;
 - a **recommended go threshold with its arithmetic**, explicitly labelled an owner-settable parameter;
 - the tumbling-window bias and its conservative direction (understates concentration).
